Updated folder structures.
